Fix Spark 3.4 build issues
