+++ pom.xml
@@ -898,6 +898,7 @@
                                 <configuration>
                                     <sources>
                                         <source>${project.basedir}/src/main/340/scala</source>
+                                        <source>${project.basedir}/src/main/340+/scala</source>
                                         <source>${project.basedir}/src/main/331+/scala</source>
                                         <source>${project.basedir}/src/main/311+-nondb/scala</source>
                                         <source>${project.basedir}/src/main/320+/java</source>
